requirements.txt
backoff==2.2.1
loguru==0.7.3
ollama==0.4.7
openai-whisper==20240930
python-dotenv==1.0.1
pytube==15.0.0
yt-dlp==2025.1.26
